backend/cart/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.2.0
pymysql
Flask-Cors
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1